+++ src/ui/layout.tsx
@@ -3,7 +3,7 @@ import { ParentComponent, Show, children, Suspense } from "solid-js";
 import { MainNavigation } from "~/ui/layout/main-navigation";
 import { MainHeader } from "./layout/main-header";
 import { Hero } from "./layout/hero";
-import { cache, createAsync, useLocation, useMatch } from "@solidjs/router";
+import { cache, createAsync, useMatch } from "@solidjs/router";
 import { DocsLayout } from "./docs-layout";
 import { PageStateProvider } from "~/data/page-state";
 import { Alert } from "@kobalte/core";
@@ -19,73 +19,9 @@ import {
 	routerEntries,
 } from "solid:collection";
 import { PathMatch } from "@solidjs/router/dist/types";
+import { useCurrentRouteMetaData } from "~/utils/route-metadata-helper";
 
 const PROJECTS = ["solid-router", "solid-start", "solid-meta"] as const;
-
-export enum Project {
-	Core = "",
-	Router = "/solid-router",
-	Start = "/solid-start",
-	Meta = "/solid-meta",
-}
-
-interface CurrentRouteMetaData {
-	project: Project | null;
-	locale: string;
-	isProjectRoot: boolean;
-}
-
-export const useCurrentRouteMetaData = (): CurrentRouteMetaData => {
-	let currentPath = useLocation().pathname;
-
-	// Trim trailing slash
-	currentPath = currentPath.endsWith("/")
-		? currentPath.slice(0, -1)
-		: currentPath;
-
-	const pathParts = currentPath.split("/").filter(Boolean);
-	const projectOrLocale: string = pathParts[0];
-
-	let returnObject: CurrentRouteMetaData = {
-		isProjectRoot: true,
-		locale: "",
-		project: null,
-	};
-
-	if (SUPPORTED_LOCALES.includes(projectOrLocale)) {
-		if (pathParts.length > 2) {
-			returnObject.isProjectRoot = false;
-		}
-
-		returnObject.locale = projectOrLocale;
-		checkPathBeyondLocale(pathParts[1] ?? "");
-	} else {
-		if (pathParts.length > 1) {
-			returnObject.isProjectRoot = false;
-		}
-
-		checkPathBeyondLocale(pathParts[0] ?? "");
-	}
-
-	function isInProjectEnum(projectPath: string): boolean {
-		return Object.values(Project).includes(projectPath as Project);
-	}
-
-	function checkPathBeyondLocale(path: string) {
-		if (path.length > 0) {
-			path = "/" + path;
-		}
-
-		if (isInProjectEnum(path)) {
-			returnObject.project = path as Project;
-		} else {
-			returnObject.isProjectRoot = false;
-			returnObject.project = "" as Project;
-		}
-	}
-
-	return returnObject;
-};
 
 function getDefaultTree(project: (typeof PROJECTS)[number]) {
 	switch (project) {
